Refactor examples
