--- examples/src/main/java/org/springframework/data/aerospike/examples/blocking/crud/MovieRepository.java
@@ -1,6 +1,0 @@
-package org.springframework.data.aerospike.examples.blocking.crud;
-
-import org.springframework.data.aerospike.repository.AerospikeRepository;
-
-public interface MovieRepository extends AerospikeRepository<MovieDocument, String> {
-}
