+++ examples/src/main/java/org/springframework/data/aerospike/examples/blocking/crud/config/AerospikeConfiguration.java
@@ -1,10 +1,13 @@
-package org.springframework.data.aerospike.examples.blocking.crud;
+package org.springframework.data.aerospike.examples.blocking.crud.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.aerospike.config.AbstractAerospikeDataConfiguration;
 import org.springframework.data.aerospike.config.AerospikeDataSettings;
+import org.springframework.data.aerospike.examples.blocking.crud.BlockingRepositoryCrudExample;
+import org.springframework.data.aerospike.examples.blocking.crud.entity.MovieDocument;
+import org.springframework.data.aerospike.examples.blocking.crud.repository.MovieRepository;
 import org.springframework.data.aerospike.repository.config.EnableAerospikeRepositories;
 
 @Configuration
