+++ examples/src/main/java/org/springframework/data/aerospike/examples/blocking/customquery/config/CustomQueryConfiguration.java
@@ -1,10 +1,13 @@
-package org.springframework.data.aerospike.examples.blocking.customquery;
+package org.springframework.data.aerospike.examples.blocking.customquery.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.aerospike.config.AbstractAerospikeDataConfiguration;
 import org.springframework.data.aerospike.config.AerospikeDataSettings;
+import org.springframework.data.aerospike.examples.blocking.customquery.CustomQueryDslExample;
+import org.springframework.data.aerospike.examples.blocking.customquery.entity.CustomQueryMovieDocument;
+import org.springframework.data.aerospike.examples.blocking.customquery.repository.CustomQueryMovieRepository;
 import org.springframework.data.aerospike.repository.config.EnableAerospikeRepositories;
 
 @Configuration
