+++ examples/src/main/java/org/springframework/data/aerospike/examples/blocking/indexed/config/IndexedAnnotationConfiguration.java
@@ -1,10 +1,13 @@
-package org.springframework.data.aerospike.examples.blocking.indexed;
+package org.springframework.data.aerospike.examples.blocking.indexed.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.aerospike.config.AbstractAerospikeDataConfiguration;
 import org.springframework.data.aerospike.config.AerospikeDataSettings;
+import org.springframework.data.aerospike.examples.blocking.indexed.IndexedAnnotationExample;
+import org.springframework.data.aerospike.examples.blocking.indexed.entity.AnnotatedMovieDocument;
+import org.springframework.data.aerospike.examples.blocking.indexed.repository.AnnotatedMovieRepository;
 import org.springframework.data.aerospike.repository.config.EnableAerospikeRepositories;
 
 @Configuration
