+++ examples/src/main/java/org/springframework/data/aerospike/examples/blocking/query/config/QueryAerospikeConfiguration.java
@@ -1,10 +1,13 @@
-package org.springframework.data.aerospike.examples.blocking.query;
+package org.springframework.data.aerospike.examples.blocking.query.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.aerospike.config.AbstractAerospikeDataConfiguration;
 import org.springframework.data.aerospike.config.AerospikeDataSettings;
+import org.springframework.data.aerospike.examples.blocking.query.SecondaryIndexQueryExample;
+import org.springframework.data.aerospike.examples.blocking.query.entity.IndexedMovieDocument;
+import org.springframework.data.aerospike.examples.blocking.query.repository.IndexedMovieRepository;
 import org.springframework.data.aerospike.repository.config.EnableAerospikeRepositories;
 
 @Configuration
