+++ examples/src/main/java/org/springframework/data/aerospike/examples/blocking/query/repository/IndexedMovieRepository.java
@@ -1,5 +1,6 @@
-package org.springframework.data.aerospike.examples.blocking.query;
+package org.springframework.data.aerospike.examples.blocking.query.repository;
 
+import org.springframework.data.aerospike.examples.blocking.query.entity.IndexedMovieDocument;
 import org.springframework.data.aerospike.repository.AerospikeRepository;
 
 import java.util.List;
