--- examples/src/main/java/org/springframework/data/aerospike/examples/reactive/crud/ReactiveMovieDocument.java
@@ -1,56 +1,0 @@
-package org.springframework.data.aerospike.examples.reactive.crud;
-
-import org.springframework.data.aerospike.mapping.Document;
-import org.springframework.data.annotation.Id;
-
-@Document(collection = "sda_examples_reactive_movies")
-public class ReactiveMovieDocument {
-
-    @Id
-    private String id;
-    private String title;
-    private int releaseYear;
-    private double rating;
-
-    public ReactiveMovieDocument() {
-    }
-
-    public ReactiveMovieDocument(String id, String title, int releaseYear, double rating) {
-        this.id = id;
-        this.title = title;
-        this.releaseYear = releaseYear;
-        this.rating = rating;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public int getReleaseYear() {
-        return releaseYear;
-    }
-
-    public void setReleaseYear(int releaseYear) {
-        this.releaseYear = releaseYear;
-    }
-
-    public double getRating() {
-        return rating;
-    }
-
-    public void setRating(double rating) {
-        this.rating = rating;
-    }
-}
