+++ examples/src/main/java/org/springframework/data/aerospike/examples/reactive/crud/ReactiveRepositoryCrudExample.java
@@ -1,12 +1,17 @@
 package org.springframework.data.aerospike.examples.reactive.crud;
 
+import org.springframework.data.aerospike.examples.reactive.crud.entity.ReactiveMovieDocument;
+import org.springframework.data.aerospike.examples.reactive.crud.repository.ReactiveMovieRepository;
 import org.springframework.stereotype.Component;
+
+import static org.springframework.data.aerospike.examples.support.ExampleAssertions.require;
 
 @Component
 public class ReactiveRepositoryCrudExample {
 
     private final ReactiveMovieRepository repository;
 
+    // Spring injects a reactive repository proxy that returns Reactor publishers.
     public ReactiveRepositoryCrudExample(ReactiveMovieRepository repository) {
         this.repository = repository;
     }
@@ -14,23 +19,27 @@ public class ReactiveRepositoryCrudExample {
     public void run() {
         ReactiveMovieDocument movie = new ReactiveMovieDocument("reactive-crud-1", "Arrival", 2016, 7.9);
 
-        repository.save(movie).block();
+        // save(...) returns a Mono; block() is used here only to keep the example sequential.
+        ReactiveMovieDocument saved = repository.save(movie).block();
 
-        ReactiveMovieDocument loaded = repository.findById(movie.getId())
+        // findById(...) emits the saved record by its @Id value.
+        ReactiveMovieDocument loaded = repository.findById(saved.getId())
             .blockOptional()
             .orElseThrow(() -> new IllegalStateException("Saved movie was not found"));
-        require("Arrival".equals(loaded.getTitle()), "Loaded movie title did not match");
-        require(Boolean.TRUE.equals(repository.existsById(movie.getId()).block()), "Saved movie should exist");
 
-        repository.deleteById(movie.getId()).block();
-        require(Boolean.FALSE.equals(repository.existsById(movie.getId()).block()), "Deleted movie should not exist");
+        String loadedTitle = loaded.getTitle();
+        require("Arrival".equals(loadedTitle), "Loaded movie title did not match");
+
+        // existsById(...) checks presence and returns the result asynchronously.
+        Boolean savedMovieExists = repository.existsById(saved.getId()).block();
+        require(Boolean.TRUE.equals(savedMovieExists), "Saved movie should exist");
+
+        // deleteById(...) removes the record and completes when the delete finishes.
+        repository.deleteById(saved.getId()).block();
+
+        Boolean deletedMovieExists = repository.existsById(saved.getId()).block();
+        require(Boolean.FALSE.equals(deletedMovieExists), "Deleted movie should not exist");
 
         System.out.println("Saved, loaded, checked, and deleted a reactive repository document");
-    }
-
-    private void require(boolean condition, String message) {
-        if (!condition) {
-            throw new IllegalStateException(message);
-        }
     }
 }
