+++ examples/src/main/java/org/springframework/data/aerospike/examples/support/ExampleAssertions.java
@@ -1,0 +1,13 @@
+package org.springframework.data.aerospike.examples.support;
+
+public final class ExampleAssertions {
+
+    private ExampleAssertions() {
+    }
+
+    public static void require(boolean condition, String message) {
+        if (!condition) {
+            throw new IllegalStateException(message);
+        }
+    }
+}
